Rename and restructure explorer node classes
