+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/explorer/PackageNodes.kt
@@ -16,37 +16,8 @@ import com.jetbrains.rider.projectView.views.*
 import com.jetbrains.rider.projectView.views.solutionExplorer.SolutionExplorerViewPane
 import icons.UnityIcons
 
-// Packages are included in a project by listing in the "dependencies" node of Packages/manifest.json. Packages can
-// contain assets, such as source, resources, and .dlls. They can also include .asmdef files
-// a) Built-in (modules). These are pseudo-package that ship built in to Unity. They can be resolved by manifests in the
-//    application directory. Modules are cached in an application specific folder, so can't be found without knowing the
-//    path to the currently running instance of Unity. Path is: Editor/Data/Resources/PackageManager/BuiltInPackages for
-//    Windows and Unity.app/Contents/Resources/PackageManager/BuiltInPackages on OSX
-// b) Referenced. These are cached in a per-user and per-registry location, and are read only. Any .asmdef files in the
-//    package will be used to compile an assembly, saved to Library/ScriptAssemblies and added as a binary reference to
-//    the project. A referenced package can be copied into the Packages folder to convert it into a writable embedded
-//    package. Cache folder is %LOCALAPPDATA%\Unity\cache\packages on Windows and ~/Library/Unity/cache/packages on OSX
-// c) Embedded. This is a package that lives inside the Packages folder, and is read-write. Any .asmdef files are used
-//    to create C# projects and added to the generated solution.
-// d) Local. These are read-write packages that live outside of the Packages folder. Any .asmdef files are used to
-//    generate C# projects. The version in Packages/manifest.json begins with `file:` and is a path to the package,
-//    either relative to the project root, or fully qualified
-// e) Git. Currently undocumented. Unity will check out a git repo to a cache folder and treat it as a read-only package
-// f) Excluded. A package can have a version of "excluded" in the manifest.json. It is simply ignored
-//
-// If there is a Packages/manifest.json file, show the Packages node in the Unity Explorer. The Packages node will show
-// all editable packages at the root, with a child node for read only packages. It will show all files and folders under
-// the Packages folder, with embedded packages being highlighted and sorted to the top. Local packages will also be
-// listed here. All other packages will be listed under "Read only", with source packages listed at the top as folders,
-// followed by modules and other/unresolved packages
-//
-// Potential actions:
-// a) If a folder has a package.json, but isn't listed in manifest.json, highlight with an error, and offer a right
-//    click action to add to manifest.json
-// b) Right click on a referenced package to convert to embedded - simply copy into the project's Packages folder
-
-class PackagesRoot(project: Project, private val packageManager: PackageManager)
-    : UnityExplorerNode(project, packageManager.packagesFolder, listOf(), AncestorNodeType.FileSystem) {
+class PackagesRootNode(project: Project, private val packageManager: PackageManager)
+    : UnityExplorerFileSystemNode(project, packageManager.packagesFolder, emptyList(), AncestorNodeType.FileSystem) {
 
     override fun update(presentation: PresentationData) {
         if (!virtualFile.isValid) return
@@ -54,27 +25,29 @@ class PackagesRoot(project: Project, private val packageManager: PackageManager)
         presentation.setIcon(UnityIcons.Explorer.PackagesRoot)
     }
 
-    override fun isAlwaysExpand() = true
-
     override fun calculateChildren(): MutableList<AbstractTreeNode<*>> {
+
+        // Add file system children, which will include embedded packages
         val children = super.calculateChildren()
 
-        // We want the children to be file system folders and editable packages, which means embedded packages and local
-        // packages. We've already added the embedded packages by including file system folders
-        packageManager.localPackages.forEach { addPackage(children, it) }
-        packageManager.unknownPackages.forEach { addPackage(children, it) }
+        // Also include any local (file: based) packages, plus all unresolved packages
+        packageManager.localPackages.forEach { addPackageNode(children, it) }
+        packageManager.unknownPackages.forEach { addPackageNode(children, it) }
 
-        if (packageManager.immutablePackages.any())
-            children.add(0, ReadOnlyPackagesRoot(project!!, packageManager))
+        // Add a root node for modules and referenced packages
+        if (packageManager.immutablePackages.any()) {
+            children.add(0, ReadOnlyPackagesRootNode(myProject, packageManager))
+        }
 
         return children
     }
 
     override fun createNode(virtualFile: VirtualFile, nestedFiles: List<NestingNode<VirtualFile>>): FileSystemNodeBase {
+        // If the child folder is an embedded package, add it as a package node
         if (virtualFile.isDirectory) {
             packageManager.getPackageData(virtualFile)?.let {
                 val embeddedPackageData = PackageData(it.name, virtualFile, it.details, PackageSource.Embedded)
-                return PackageNode(project!!, packageManager, virtualFile, embeddedPackageData)
+                return PackageNode(myProject, packageManager, virtualFile, embeddedPackageData)
             }
         }
         return super.createNode(virtualFile, nestedFiles)
@@ -84,18 +57,18 @@ class PackagesRoot(project: Project, private val packageManager: PackageManager)
     // True is effectively "maybe"
     override fun contains(file: VirtualFile) = true
 
-    private fun addPackage(children: MutableList<AbstractTreeNode<*>>, thePackage: PackageData) {
+    private fun addPackageNode(children: MutableList<AbstractTreeNode<*>>, thePackage: PackageData) {
         if (thePackage.packageFolder != null) {
-            children.add(PackageNode(project!!, packageManager, thePackage.packageFolder, thePackage))
+            children.add(PackageNode(myProject, packageManager, thePackage.packageFolder, thePackage))
         }
         else {
-            children.add(UnknownPackageNode(project!!, thePackage))
+            children.add(UnknownPackageNode(myProject, thePackage))
         }
     }
 }
 
 class PackageNode(project: Project, private val packageManager: PackageManager, packageFolder: VirtualFile, private val packageData: PackageData)
-    : UnityExplorerNode(project, packageFolder, listOf(), AncestorNodeType.fromPackageData(packageData)), Comparable<AbstractTreeNode<*>> {
+    : UnityExplorerFileSystemNode(project, packageFolder, emptyList(), AncestorNodeType.fromPackageData(packageData)), Comparable<AbstractTreeNode<*>> {
 
     init {
         icon = when (packageData.source) {
@@ -117,8 +90,9 @@ class PackageNode(project: Project, private val packageManager: PackageManager, 
         presentation.addNonIndexedMark(myProject, virtualFile)
 
         // Note that this might also set the tooltip if we have too many projects underneath
-        if (UnityExplorer.getInstance(myProject).showProjectNames)
+        if (UnityExplorer.getInstance(myProject).showProjectNames) {
             addProjects(presentation)
+        }
 
         val existingTooltip = presentation.tooltip ?: ""
 
@@ -150,7 +124,7 @@ class PackageNode(project: Project, private val packageManager: PackageManager, 
         val children = super.calculateChildren()
 
         if (packageData.details.dependencies.isNotEmpty()) {
-            children.add(0, DependenciesRoot(project!!, packageManager, packageData))
+            children.add(0, PackageDependenciesRoot(myProject, packageManager, packageData))
         }
 
         return children
@@ -162,7 +136,7 @@ class PackageNode(project: Project, private val packageManager: PackageManager, 
     }
 }
 
-class DependenciesRoot(project: Project, private val packageManager: PackageManager, private val packageData: PackageData)
+class PackageDependenciesRoot(project: Project, private val packageManager: PackageManager, private val packageData: PackageData)
     : AbstractTreeNode<Any>(project, packageData) {
 
     override fun update(presentation: PresentationData) {
@@ -173,13 +147,13 @@ class DependenciesRoot(project: Project, private val packageManager: PackageMana
     override fun getChildren(): MutableCollection<AbstractTreeNode<*>> {
         val children = mutableListOf<AbstractTreeNode<*>>()
         for ((name, version) in packageData.details.dependencies) {
-            children.add(DependencyItemNode(project!!, packageManager, name, version))
+            children.add(PackageDependencyItemNode(myProject, packageManager, name, version))
         }
         return children
     }
 }
 
-class DependencyItemNode(project: Project, private val packageManager: PackageManager, private val packageName: String, version: String)
+class PackageDependencyItemNode(project: Project, private val packageManager: PackageManager, private val packageName: String, version: String)
     : AbstractTreeNode<Any>(project, "$packageName@$version") {
 
     init {
@@ -201,31 +175,31 @@ class DependencyItemNode(project: Project, private val packageManager: PackageMa
     override fun navigate(requestFocus: Boolean) {
         val packageData = packageManager.getPackageData(packageName)
         if (packageData?.packageFolder == null) return
-        project!!.navigateToSolutionView(packageData.packageFolder, requestFocus)
+        myProject.navigateToSolutionView(packageData.packageFolder, requestFocus)
     }
 }
 
-abstract class CompositeFolderRoot(project: Project, key: Any)
+abstract class FolderContainerNodeBase(project: Project, key: Any)
     : SolutionViewNode<Any>(project, key) {
 
-    private val packageFolders = mutableSetOf<VirtualFile>()
+    private val childFolders = mutableSetOf<VirtualFile>()
 
-    protected fun addPackageFolder(packageFolder: VirtualFile) {
-        packageFolders.add(packageFolder)
+    protected fun addChildFolder(packageFolder: VirtualFile) {
+        childFolders.add(packageFolder)
     }
 
     // Note that this requires the children to have been expanded first. The SolutionViewVisitor will ensure this happens
     override fun contains(file: VirtualFile): Boolean {
-        if (packageFolders.contains(file)) return true
-        for (packageFolder in packageFolders) {
+        if (childFolders.contains(file)) return true
+        for (packageFolder in childFolders) {
             if (VfsUtil.isAncestor(packageFolder,  file, false)) return true
         }
         return false
     }
 }
 
-class ReadOnlyPackagesRoot(project: Project, private val packageManager: PackageManager)
-    : CompositeFolderRoot(project, key) {
+class ReadOnlyPackagesRootNode(project: Project, private val packageManager: PackageManager)
+    : FolderContainerNodeBase(project, key) {
 
     companion object {
         val key = Any()
@@ -240,15 +214,15 @@ class ReadOnlyPackagesRoot(project: Project, private val packageManager: Package
         val children = mutableListOf<AbstractTreeNode<*>>()
 
         if (packageManager.hasBuiltInPackages) {
-            children.add(BuiltinPackagesRoot(project!!, packageManager))
+            children.add(BuiltinPackagesRootNode(myProject, packageManager))
 
             // Add the builtin packages root folder to the list of folders we know are under this node. This lets us
             // correctly handle `contains()` for built in packages, which in turn means we can navigate to a built in
             // package by folder (e.g. by double clicking a dependency node)
             try {
-                UnityInstallationFinder.getInstance(myProject).getBuiltInPackagesRoot()?.let {
-                    VfsUtil.findFile(it, true)?.let { vfile ->
-                        addPackageFolder(vfile)
+                UnityInstallationFinder.getInstance(myProject).getBuiltInPackagesRoot()?.let { path ->
+                    VfsUtil.findFile(path, true)?.let {
+                        addChildFolder(it)
                     }
                 }
             } catch (throwable: Throwable) {
@@ -260,19 +234,19 @@ class ReadOnlyPackagesRoot(project: Project, private val packageManager: Package
             if (packageData.source == PackageSource.BuiltIn) continue
 
             if (packageData.packageFolder == null) {
-                children.add(UnknownPackageNode(project!!, packageData))
+                children.add(UnknownPackageNode(myProject, packageData))
             }
             else {
-                children.add(PackageNode(project!!, packageManager, packageData.packageFolder, packageData))
-                addPackageFolder(packageData.packageFolder)
+                children.add(PackageNode(myProject, packageManager, packageData.packageFolder, packageData))
+                addChildFolder(packageData.packageFolder)
             }
         }
         return children
     }
 }
 
-class BuiltinPackagesRoot(project: Project, private val packageManager: PackageManager)
-    : CompositeFolderRoot(project, key) {
+class BuiltinPackagesRootNode(project: Project, private val packageManager: PackageManager)
+    : FolderContainerNodeBase(project, key) {
 
     companion object {
         val key = Any()
@@ -289,11 +263,11 @@ class BuiltinPackagesRoot(project: Project, private val packageManager: PackageM
             if (packageData.source != PackageSource.BuiltIn) continue
 
             if (packageData.packageFolder == null) {
-                children.add(UnknownPackageNode(project!!, packageData))
+                children.add(UnknownPackageNode(myProject, packageData))
             }
             else {
-                children.add(BuiltinPackageNode(project!!, packageData))
-                addPackageFolder(packageData.packageFolder)
+                children.add(BuiltinPackageNode(myProject, packageData))
+                addChildFolder(packageData.packageFolder)
             }
         }
         return children
@@ -305,7 +279,7 @@ class BuiltinPackagesRoot(project: Project, private val packageManager: PackageM
 // Note that a module can have dependencies. Perhaps we want to always show this as a folder, including the Dependencies
 // node?
 class BuiltinPackageNode(project: Project, private val packageData: PackageData)
-    : UnityExplorerNode(project, packageData.packageFolder!!, listOf(), AncestorNodeType.ReadOnlyPackage) {
+    : UnityExplorerFileSystemNode(project, packageData.packageFolder!!, emptyList(), AncestorNodeType.ReadOnlyPackage) {
 
     override fun calculateChildren(): MutableList<AbstractTreeNode<*>> {
 
@@ -325,7 +299,7 @@ class BuiltinPackageNode(project: Project, private val packageData: PackageData)
     }
 
     override fun createNode(virtualFile: VirtualFile, nestedFiles: List<NestingNode<VirtualFile>>): FileSystemNodeBase {
-        return UnityExplorerNode(project!!, virtualFile, nestedFiles, descendentOf)
+        return UnityExplorerFileSystemNode(myProject, virtualFile, nestedFiles, descendentOf)
     }
 
     override fun canNavigateToSource(): Boolean {
@@ -342,7 +316,7 @@ class BuiltinPackageNode(project: Project, private val packageData: PackageData)
 
         val packageJson = virtualFile.findChild("package.json")
         if (packageJson != null) {
-            OpenFileDescriptor(project!!, packageJson).navigate(requestFocus)
+            OpenFileDescriptor(myProject, packageJson).navigate(requestFocus)
         }
     }
 
@@ -362,7 +336,7 @@ class BuiltinPackageNode(project: Project, private val packageData: PackageData)
     }
 }
 
-// Note that this might get a PackageData with source == PackageSource.Unknown
+// Note that this might get a PackageData with source == PackageSource.BuiltIn
 class UnknownPackageNode(project: Project, private val packageData: PackageData)
     : AbstractTreeNode<Any>(project, packageData) {
 
@@ -380,6 +354,8 @@ class UnknownPackageNode(project: Project, private val packageData: PackageData)
     override fun update(presentation: PresentationData) {
         presentation.addText(name, SimpleTextAttributes.REGULAR_ATTRIBUTES)
         presentation.setIcon(icon)
+
+        // Description can be an error message
         if (packageData.details.description.isNotEmpty()) {
             presentation.tooltip = formatDescription(packageData)
         }
