+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/explorer/UnityExplorer.kt
@@ -17,7 +17,7 @@ import com.jetbrains.rider.projectView.views.solutionExplorer.SolutionExplorerVi
 import icons.UnityIcons
 import org.jdom.Element
 
-class UnityExplorer(project: Project) : SolutionViewPaneBase(project, UnityExplorerRootNode(project, PackageManager.getInstance(project))) {
+class UnityExplorer(project: Project) : SolutionViewPaneBase(project, createRootNode(project)) {
 
     companion object {
         const val ID = "UnityExplorer"
@@ -34,6 +34,10 @@ class UnityExplorer(project: Project) : SolutionViewPaneBase(project, UnityExplo
 
         fun tryGetInstance(project: Project): UnityExplorer? {
             return ProjectView.getInstance(project).getProjectViewPaneById(ID) as? UnityExplorer
+        }
+
+        private fun createRootNode(project: Project): UnityExplorerRootNode {
+            return UnityExplorerRootNode(project, PackageManager.getInstance(project))
         }
     }
 
@@ -53,7 +57,7 @@ class UnityExplorer(project: Project) : SolutionViewPaneBase(project, UnityExplo
         val root = tree.model.root
         val count = tree.model.getChildCount(root)
         for (i in 0..count) {
-            if (tree.model.getChild(root, i) is PackagesRoot) {
+            if (tree.model.getChild(root, i) is PackagesRootNode) {
                 return true
             }
         }
