+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/explorer/UnityProjectModelViewExtensions.kt
@@ -10,6 +10,7 @@ import com.jetbrains.rider.projectView.ProjectEntityView
 import com.jetbrains.rider.projectView.ProjectModelViewExtensions
 import com.jetbrains.rider.projectView.workspace.*
 
+@Suppress("UnstableApiUsage")
 class UnityProjectModelViewExtensions(project: Project) : ProjectModelViewExtensions(project) {
 
     // this is called for rename, we should filter .Player projects and return node itself
